chore(clerk-js): Update snapshots to omit the deprecated class properties
The `logoUrl` properties from the snapshots even though they still
exist in those classes because of the way the deprecation helper
works to mark the property as deprecated.
The deprecation helper redefine the property in the class prototype
using a getter/setter that shows a warning in getter using the
Object.defineProperty(). By redefining a property, that property
is not considered class own property any more which causes the
jest snapshot and the JSON.stringify() to omit that property.
This is a side-effect of the deprecation implementation and not a
spec but i think that we are okay to proceed with omitting the
properties from snapshots.